requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
boto3>=1.34.0
botocore>=1.34.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pdfplumber>=0.9.0
openpyxl
pytesseract
pillow>=9.5.0
beautifulsoup4
apscheduler>=3.10.0